feat: kick player out of a game
When host clicks on the cross icon of a player, after confirmation, player socket is removed from game. Kicked player is notified and redirected to home.
Game_state is actualized via removePlayer function.
Choice has been made to reset Localstorage data from Kicked-player to not cause any localStorage issues if player wants to join or create another game.
Limits: each time a player leaves or is kicked out of a game, he has to enter a new pseudo and create a new socket to create or join a new game.